# public/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "Taklimakan cryptodata",
    "version": "v0.0.1"
  },
  "paths": {
...
    "/api/top/tn_indexes/charts": {
      "get": {
        "tags": [
          "top"
        ],
        "summary": "Get data for charts TN index",
        "description": "Data charts interval time from to",
        "operationId": "getCharts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Index type",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "tn10",
                "tn50",
                "tn100",
                "tn200"
              ]
            },
            "collectionFormat": "multi"
          },
          {
            "name": "period_date_start",
            "in": "query",
            "description": "Date start",
            "required": false,
            "type": "string"
          },
          {
            "name": "period_date_end",
            "in": "query",
            "description": "Date end",
            "required": false,
            "type": "string"
          },
          {
            "name": "object_amount",
            "in": "query",
            "description": "Amount object in response",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "photoUrls"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "period_date_end": {
                      "type": "string"
                    },
                    "period_interval": {
                      "type": "string"
                    },
                    "object_amount": {
                      "type": "string"
                    }
                  }
                },
                "data": {
                  "type": "array",
                  "xml": {
                    "name": "",
                    "wrapped": true
                  },
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/top/tn_indexes/get": {
      "get": {
        "tags": [
          "top"
        ],
        "summary": "Get list tn indexes",
        "description": "list tn indexes with coin list",
        "operationId": "getIndexes",
        "produces": [
          "application/json"
        ],
        "parameters": null,
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "photoUrls"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "data": {
                  "type": "array",
                  "xml": {
                    "name": "",
                    "wrapped": true
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "default": {
                        "type": "boolean"
                      },
                      "count_coins": {
                        "type": "integer"
                      },
                      "coins": {
                        "type": "array",
                        "xml": {
                          "name": "",
                          "wrapped": true
                        },
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/top/tn_indexes/new": {
      "post": {
        "tags": [
          "top"
        ],
        "summary": "new index",
        "description": "Create new index",
        "operationId": "newIndex",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Index name",
            "required": true,
            "type": "string"
          },
          {
            "name": "coin_list",
            "in": "query",
            "description": "List coins",
            "required": true,
            "type": "string"
          },
          {
            "name": "default",
            "in": "query",
            "description": "Default or custom index",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "photoUrls"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "coin_list": {
                      "type": "string"
                    },
                    "default": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/top/tn_indexes/change": {
      "post": {
        "tags": [
          "top"
        ],
        "summary": "Change index",
        "description": "Change index",
        "operationId": "changeIndex",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Index name",
            "required": true,
            "type": "string"
          },
          {
            "name": "coin_list",
            "in": "query",
            "description": "List coins",
            "required": true,
            "type": "string"
          },
          {
            "name": "default",
            "in": "query",
            "description": "Default or custom index",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "photoUrls"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "coin_list": {
                      "type": "string"
                    },
                    "default": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/top/tn_indexes/delete": {
      "post": {
        "tags": [
          "top"
        ],
        "summary": "delete index",
        "description": "delete index",
        "operationId": "deleteIndex",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Index name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "photoUrls"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
      },
    "consumes": [
      "application/json"
    ]
  }